+++ src/base/providers/ExceptionProvider.ts
@@ -1,6 +1,5 @@
 import { DdRum, DdSdkReactNative } from '@datadog/mobile-react-native';
 import { navigationRef } from '../../utils/navigationRef';
-import Sentry from '../../config/sentryConfig';
 
 export interface IUser {
   id: string;
@@ -12,40 +11,6 @@ export class ExceptionProvider {
   private static oldRouteName: string = '';
 
   private static oldRouteParams: Readonly<object | undefined>;
-
-  static captureException(
-    error: Error,
-    params?: { [key: string]: unknown },
-    tags?: { [key: string]: unknown },
-    breadcrumbs?: { [key: string]: unknown },
-  ) {
-    if (!params) {
-      // Sentry.captureException(error);
-      // return;
-    }
-
-    // Sentry.withScope((scope) => {
-    //   Object.keys(params).forEach((key) => {
-    //     scope.setExtra(key, params[key]);
-    //   });
-
-    //   if (tags && Object.keys(tags).length) {
-    //     Object.keys(tags).forEach((key) => {
-    //       // @ts-ignore
-    //       scope.setTag(key, tags[key]);
-    //     });
-    //   }
-
-    //   if (breadcrumbs && Object.keys(breadcrumbs).length) {
-    //     Object.keys(breadcrumbs).forEach((key) => {
-    //       // @ts-ignore
-    //       scope.addBreadcrumb(key, breadcrumbs[key]);
-    //     });
-    //   }
-
-    //   Sentry.captureException(error);
-    // });
-  }
 
   static trackScreen() {
     try {
@@ -63,8 +28,6 @@ export class ExceptionProvider {
       this.oldRouteName = screenName;
       this.oldRouteParams = params;
 
-      // Sentry.addBreadcrumb({ message: screenName, category: 'navigation', data: params });
-
       DdRum.startView(
         screenName,
         screenName,
@@ -72,7 +35,7 @@ export class ExceptionProvider {
         Date.now(),
       );
     } catch (e) {
-      this.captureException(e);
+      console.log(e);
     }
   }
 
@@ -82,14 +45,9 @@ export class ExceptionProvider {
       name: user.name,
       email: user.email,
     });
-
-    // Sentry.configureScope((scope) => {
-    //   scope.setUser({ email: user.email });
-    // });
   }
 
   static unsetUser() {
     DdSdkReactNative.setUser({});
-    // Sentry.setUser(null);
   }
 }