+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 2.26.5
 - [Implementação evento Revenue, Average order value e Purchase Rate - Algolia](https://dev.azure.com/ARECO/AReCO/_workitems/edit/55205)
 - [Preço do produto zerado nos favoritos](https://dev.azure.com/ARECO/AReCO/_workitems/edit/56038)
+- [Remoção do Sentry no projeto](Sem url)
 
 
 ### 2.26.3